+++ d2/src/main/java/com/linkedin/d2/jmx/NoOpXdsClientOtelMetricsProvider.java
@@ -93,4 +93,12 @@ public class NoOpXdsClientOtelMetricsProvider implements XdsClientOtelMetricsPro
   public void updateActiveInitialWaitTime(String clientName, long waitTimeMs) {
     // No-op
   }
+
+  /**
+   * {@inheritDoc}
+   */
+  @Override
+  public void recordRequestSizeBytes(String clientName, long sizeBytes) {
+    // No-op
+  }
 }